refactor: reorganize CLI code into dedicated package
- Split CLI code into dedicated package structure
- Create separate modules for commands, formatters, and exporters
- Improve error handling for API key configuration
- Simplify main entry point
- Update changelog
