+++ CHANGELOG.md
@@ -2,74 +2,57 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.5.0] - 2024-10-22
-
-### Added
-- Enhanced search functionality:
-  - Search by tournament slug to find organizer
-  - Get tournament organizer information
-  - Interactive tournament selection
-- Custom GraphQL query for tournament owner information
-- Improved error handling for search operations
-- Updated documentation with new search features
+## [1.4.0] - 2024-10-23
 
 ### Changed
-- Search command now supports both owner ID and tournament slug
-- Improved search results display with more tournament details
-- Enhanced help text with more examples and usage information
+- Refactored CLI code into a dedicated package structure for better maintainability
+  - Separated commands into individual modules (search.py, results.py)
+  - Created dedicated formatters for table display
+  - Isolated export functionality
+  - Improved error handling for API key configuration
+- Improved code organization and separation of concerns
+- Simplified main entry point (startgg.py)
 
-## [1.4.0] - 2024-01-14
-
-### Added
-- New CLI interface using Typer and Rich libraries
-- Search command to find tournaments by owner ID
-- Improved formatting with colored output and tables
-- Progress indicators for long-running operations
-- Better help text and command documentation
-
-### Changed
-- Renamed main script from app.py to startgg.py
-- Restructured commands into a more organized CLI
-- Updated documentation with new command structure
-- Improved error handling and user feedback
-
-### Removed
-- Old app.py script in favor of new CLI structure
-
-## [1.3.0] - 2024-01-10
+## [1.3.0] - 2024-09-15
 
 ### Added
-- Export functionality for results in JSON, CSV, and TXT formats
-- Command-line flags for export options (-j, -c, -t)
-- Detailed help information accessible via -h flag
-- Display of total number of participants for each event
+- Added support for searching tournaments by owner ID
+- Added interactive tournament selection with --select flag
+- Added CSV export functionality for tournament results
+- Added TXT export functionality for tournament results
 
 ### Changed
-- Improved tournament information display format
-- Replace tournament slug with full tournament name in output
-- Enhanced readability of Top 8 results output
+- Improved error handling for API requests
+- Enhanced tournament information display
+- Updated documentation with new features
 
-### Fixed
-- Error handling for API responses
-
-## [1.2.0] - 2023-06-01
+## [1.2.0] - 2024-08-20
 
 ### Added
-- Support for fetching results from multiple events in a tournament
+- Added JSON export functionality for tournament results
+- Added support for multiple events in tournaments
+- Added tournament date range display
 
 ### Changed
-- Improved error handling and user feedback
+- Improved table formatting for better readability
+- Enhanced error messages for better user feedback
 
-## [1.1.0] - 2023-05-15
+## [1.1.0] - 2024-07-10
 
 ### Added
-- Command-line interface for specifying tournament slug
+- Added support for tournament search by slug
+- Added rich table display for tournament results
+- Added pagination support for tournament listings
 
 ### Changed
-- Refactored code for better modularity
+- Improved command-line interface
+- Enhanced error handling
+- Updated documentation
 
-## [1.0.0] - 2023-05-01
+## [1.0.0] - 2024-06-01
 
 ### Added
 - Initial release
-- Basic functionality to fetch Top 8 results from a Smash.gg tournament
+- Basic tournament results fetching
+- Command-line interface
+- Tournament search functionality
